+++ extensions/vscode/src/utils/errorTypes.test.ts
@@ -271,13 +271,16 @@ describe("ErrTOMLValidationError", () => {
         line: 7,
         column: 1,
         key: "shortcut_key",
+        message: "foo: must be an integer",
       },
     });
 
     const msg = errTOMLValidationErrorMessage(
       err as axiosErrorWithJson<ErrTOMLValidationError>,
     );
-    expect(msg).toBe(`The Configuration has a schema error`);
+    expect(msg).toBe(
+      `The Configuration has a schema error: foo: must be an integer`,
+    );
   });
 });