+++ extensions/vscode/src/views/homeView.ts
@@ -37,6 +37,8 @@ import {
   AllContentRecordTypes,
   EnvironmentConfig,
   PreContentRecordWithConfig,
+  ProductName,
+  ServerType,
 } from "src/api";
 import { EventStream } from "src/events";
 import { getPythonInterpreterPath, getRInterpreterPath } from "../utils/vscode";
@@ -94,7 +96,12 @@ import {
   SelectionIsPreContentRecord,
   setSelectionIsPreContentRecord,
 } from "../extension";
-import { createNewCredentialLabel } from "src/multiStepInputs/common";
+import {
+  createNewCredentialLabel,
+  getProductType,
+  isConnectCloudProduct,
+  isConnectProduct,
+} from "src/utils/multiStepHelpers";
 
 enum HomeViewInitialized {
   initialized = "initialized",
@@ -1234,7 +1241,15 @@ export class HomeViewProvider implements WebviewViewProvider, Disposable {
         if (credential?.name) {
           details.push(credential.name);
         } else {
-          details.push(`Missing Credential for ${contentRecord.serverUrl}`);
+          const serverType = contentRecord.serverType || ServerType.CONNECT;
+          const productType = getProductType(serverType);
+          let message = "Missing Credential";
+          if (isConnectCloudProduct(productType)) {
+            message += ` for ${ProductName.CONNECT_CLOUD}`;
+          } else if (isConnectProduct(productType)) {
+            message += ` for ${contentRecord.serverUrl}`;
+          }
+          details.push(message);
           problem = true;
         }
 
@@ -1811,7 +1826,7 @@ export class HomeViewProvider implements WebviewViewProvider, Disposable {
     this.configWatchers?.dispose();
   }
 
-  public register(watchers: WatcherManager) {
+  public async register(watchers: WatcherManager) {
     this.stream.register("publish/start", () => {
       this.onPublishStart();
     });
@@ -1830,6 +1845,22 @@ export class HomeViewProvider implements WebviewViewProvider, Disposable {
       }),
     );
 
+    const currentContentRecord = await this.state.getSelectedContentRecord();
+    const serverType = currentContentRecord?.serverType || ServerType.CONNECT;
+    const productType = getProductType(serverType);
+
+    // register this command for Connect only,
+    // since Connect Cloud does not support this feature at the moment
+    if (isConnectProduct(productType)) {
+      this.context.subscriptions.push(
+        commands.registerCommand(
+          Commands.HomeView.AssociateDeployment,
+          () => showAssociateGUID(this.state),
+          this,
+        ),
+      );
+    }
+
     this.context.subscriptions.push(
       commands.registerCommand(
         Commands.HomeView.SelectDeployment,
@@ -1847,11 +1878,6 @@ export class HomeViewProvider implements WebviewViewProvider, Disposable {
       commands.registerCommand(
         Commands.HomeView.ShowSelectConfigForDeployment,
         this.showSelectOrCreateConfigForDeployment,
-        this,
-      ),
-      commands.registerCommand(
-        Commands.HomeView.AssociateDeployment,
-        () => showAssociateGUID(this.state),
         this,
       ),
       commands.registerCommand(